refactor: console 코드 삭제
