+++ src/test/java/roomescape/RoomescapeApplicationTest.java
@@ -3,9 +3,9 @@ package roomescape;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class RoomescapeApplicationTest {
-        @Test
-        void contextLoads() {
-        }
+    @Test
+    void contextLoads() {
+    }
 }
