+++ src/test/java/roomescape/dao/ReservationTimeDaoTest.java
@@ -8,25 +8,47 @@ import static roomescape.TestFixture.RESERVATION_TIME_FIXTURE;
 import static roomescape.TestFixture.ROOM_THEME_FIXTURE;
 import static roomescape.TestFixture.TIME_FIXTURE;
 
+import io.restassured.RestAssured;
+import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.dao.EmptyResultDataAccessException;
-import roomescape.console.dao.InMemoryReservationTimeDao;
-import roomescape.console.db.InMemoryReservationDb;
-import roomescape.console.db.InMemoryReservationTimeDb;
 import roomescape.domain.Name;
 import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
+import roomescape.domain.RoomTheme;
 
-class InMemoryReservationTimeDaoTest {
-    private InMemoryReservationDb inMemoryReservationDb;
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class ReservationTimeDaoTest {
+    @LocalServerPort
+    private int port;
+
+    @Autowired
+    private ReservationDao reservationDao;
+    @Autowired
     private ReservationTimeDao reservationTimeDao;
+    @Autowired
+    private RoomThemeDao roomThemeDao;
 
     @BeforeEach
     void setUp() {
-        inMemoryReservationDb = new InMemoryReservationDb();
-        reservationTimeDao = new InMemoryReservationTimeDao(
-                inMemoryReservationDb, new InMemoryReservationTimeDb());
+        RestAssured.port = port;
+        List<Reservation> reservations = reservationDao.findAll();
+        for (Reservation reservation : reservations) {
+            reservationDao.deleteById(reservation.getId());
+        }
+        List<ReservationTime> reservationTimes = reservationTimeDao.findAll();
+        for (ReservationTime reservationTime : reservationTimes) {
+            reservationTimeDao.deleteById(reservationTime.getId());
+        }
+        List<RoomTheme> roomThemes = roomThemeDao.findAll();
+        for (RoomTheme roomTheme : roomThemes) {
+            roomThemeDao.deleteById(roomTheme.getId());
+        }
     }
 
     @DisplayName("모든 예약 시간을 보여준다")
@@ -48,9 +70,10 @@ class InMemoryReservationTimeDaoTest {
     @Test
     void findById() {
         // given & when
-        reservationTimeDao.save(RESERVATION_TIME_FIXTURE);
+        ReservationTime reservationTime = reservationTimeDao.save(RESERVATION_TIME_FIXTURE);
+        Long id = reservationTime.getId();
         // then
-        assertThat(reservationTimeDao.findById(1L).getStartAt()).isEqualTo(TIME_FIXTURE);
+        assertThat(reservationTimeDao.findById(id).getStartAt()).isEqualTo(TIME_FIXTURE);
     }
 
     @DisplayName("해당 id의 예약 시간이 없는 경우, 예외가 발생한다.")
@@ -79,9 +102,9 @@ class InMemoryReservationTimeDaoTest {
     @Test
     void deleteById() {
         // given
-        reservationTimeDao.save(RESERVATION_TIME_FIXTURE);
+        ReservationTime reservationTime = reservationTimeDao.save(RESERVATION_TIME_FIXTURE);
         // when
-        reservationTimeDao.deleteById(1L);
+        reservationTimeDao.deleteById(reservationTime.getId());
         // then
         assertThat(reservationTimeDao.findAll()).isEmpty();
     }
@@ -90,22 +113,23 @@ class InMemoryReservationTimeDaoTest {
     @Test
     void deleteByIdDeletesReservationAlso() {
         // given
-        reservationTimeDao.save(RESERVATION_TIME_FIXTURE);
-        inMemoryReservationDb.insert(new Reservation(new Name("aa"), DATE_FIXTURE,
-                RESERVATION_TIME_FIXTURE, ROOM_THEME_FIXTURE));
+        ReservationTime reservationTime = reservationTimeDao.save(RESERVATION_TIME_FIXTURE);
+        RoomTheme roomTheme = roomThemeDao.save(ROOM_THEME_FIXTURE);
+        reservationDao.save(new Reservation(new Name("aa"), DATE_FIXTURE,
+                reservationTime, roomTheme));
         // when
-        reservationTimeDao.deleteById(1L);
+        reservationTimeDao.deleteById(reservationTime.getId());
         // then
-        assertThat(inMemoryReservationDb.selectAll()).isEmpty();
+        assertThat(reservationDao.findAll()).isEmpty();
     }
 
     @DisplayName("삭제 대상이 존재하면 true를 반환한다.")
     @Test
     void returnTrueWhenDeleted() {
         // given
-        reservationTimeDao.save(RESERVATION_TIME_FIXTURE);
+        ReservationTime reservationTime = reservationTimeDao.save(RESERVATION_TIME_FIXTURE);
         // when
-        boolean deleted = reservationTimeDao.deleteById(1L);
+        boolean deleted = reservationTimeDao.deleteById(reservationTime.getId());
         // then
         assertThat(deleted).isTrue();
     }
